requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
typing_extensions==4.12.2
urllib3==2.3.0
